requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
python-dotenv
langchain-community
streamlit
voyageai
pydantic
langchain-anthropic
langchain-pinecone
PyPDF2